Se implementan likes a los comentarios

+++ database/migrations/2019_12_11_003940_create_comments_table.php
@@ -15,8 +15,8 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('status_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('status_id');
             $table->text('body');
             $table->timestamps();
